requirements.txt
crewai>=0.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
vinted_scraper>=0.1.0
jinja2>=3.0.0
rich>=13.0.0
colorama>=0.4.6
google-generativeai>=0.3.0
crewai_tools>=0.1.0

